+++ src/components/Cypress/TestData.js
@@ -28,6 +28,14 @@ class TestData {
 
     /**
      *
+     * @returns {*}
+     */
+    getState() {
+        return this._state;
+    }
+
+    /**
+     *
      * @returns {boolean}
      */
     isPassed() {
@@ -36,10 +44,18 @@ class TestData {
 
     /**
      *
-     * @returns {*}
+     * @returns {boolean}
      */
-    getState() {
-        return this._state;
+    isFailed() {
+        return this._state === 'failed';
+    }
+
+    /**
+     *
+     * @returns {boolean}
+     */
+    isSkipped() {
+        return this._state === 'pending';
     }
 
     /**