+++ src/components/TestRail/Result.js
@@ -13,6 +13,10 @@ class Result {
         this._comment = comment;
         this._durationMS = duration;
         this._screenshotPaths = screenshotPaths;
+
+        if (this._screenshotPaths === undefined) {
+            this._screenshotPaths = [];
+        }
     }
 
     /**